Time: 5 ms (31.59%), Space: 69.5 MB (63.69%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -12,6 +12,6 @@ class Solution {
             sum=sum+nums[i]-nums[i-k];
             max1=Math.max(max1,sum);
         }
-        return max1/k;
+        return (double) max1/k;
     }
 }
